// src/api/modules/anilist/graphql/generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** Internal type. DO NOT USE DIRECTLY. */
export type Incremental<T> =
  | T
  | { [P in keyof T]?: P extends " $fragmentName" | "__typename" ? T[P] : never };
import type { DocumentNode } from "graphql";
export type ExternalLinkType = "INFO" | "SOCIAL" | "STREAMING";

/** The format the media was released in */
export type MediaFormat =
  /** Professionally published manga with more than one chapter */
  | "MANGA"
  /** Anime movies with a theatrical release */
  | "MOVIE"
  /** Short anime released as a music video */
  | "MUSIC"
  /** Written books released as a series of light novels */
  | "NOVEL"
  /** (Original Net Animation) Anime that have been originally released online or are only available through streaming services. */
  | "ONA"
  /** Manga with just one chapter */
  | "ONE_SHOT"
  /** (Original Video Animation) Anime that have been released directly on DVD/Blu-ray without originally going through a theatrical release or television broadcast */
  | "OVA"
  /** Special episodes that have been included in DVD/Blu-ray releases, picture dramas, pilots, etc */
  | "SPECIAL"
  /** Anime broadcast on television */
  | "TV"
  /** Anime which are under 15 minutes in length and broadcast on television */
  | "TV_SHORT";

/** Media list watching/reading status enum. */
export type MediaListStatus =
  /** Finished watching/reading */
  | "COMPLETED"
  /** Currently watching/reading */
  | "CURRENT"
  /** Stopped watching/reading before completing */
  | "DROPPED"
  /** Paused watching/reading */
  | "PAUSED"
  /** Planning to watch/read */
  | "PLANNING"
  /** Re-watching/reading */
  | "REPEATING";

export type MediaSeason =
  /** Predominantly started airing between October and November */
  | "FALL"
  /** Predominantly started airing between April and June */
  | "SPRING"
  /** Predominantly started airing between July and September */
  | "SUMMER"
  /** Predominantly started airing between January and March */
  | "WINTER";

/** Media sort enums */
export type MediaSort =
  | "CHAPTERS"
  | "CHAPTERS_DESC"
  | "DURATION"
  | "DURATION_DESC"
  | "END_DATE"
  | "END_DATE_DESC"
  | "EPISODES"
  | "EPISODES_DESC"
  | "FAVOURITES"
  | "FAVOURITES_DESC"
  | "FORMAT"
  | "FORMAT_DESC"
  | "ID"
  | "ID_DESC"
  | "POPULARITY"
  | "POPULARITY_DESC"
  | "SCORE"
  | "SCORE_DESC"
  | "SEARCH_MATCH"
  | "START_DATE"
  | "START_DATE_DESC"
  | "STATUS"
  | "STATUS_DESC"
  | "TITLE_ENGLISH"
  | "TITLE_ENGLISH_DESC"
  | "TITLE_NATIVE"
  | "TITLE_NATIVE_DESC"
  | "TITLE_ROMAJI"
  | "TITLE_ROMAJI_DESC"
  | "TRENDING"
  | "TRENDING_DESC"
  | "TYPE"
  | "TYPE_DESC"
  | "UPDATED_AT"
  | "UPDATED_AT_DESC"
  | "VOLUMES"
  | "VOLUMES_DESC";

/** Source type the media was adapted from */
export type MediaSource =
  /** Version 2+ only. Japanese Anime */
  | "ANIME"
  /** Version 3 only. Comics excluding manga */
  | "COMIC"
  /** Version 2+ only. Self-published works */
  | "DOUJINSHI"
  /** Version 3 only. Games excluding video games */
  | "GAME"
  /** Written work published in volumes */
  | "LIGHT_NOVEL"
  /** Version 3 only. Live action media such as movies or TV show */
  | "LIVE_ACTION"
  /** Asian comic book */
  | "MANGA"
  /** Version 3 only. Multimedia project */
  | "MULTIMEDIA_PROJECT"
  /** Version 2+ only. Written works not published in volumes */
  | "NOVEL"
  /** An original production not based of another work */
  | "ORIGINAL"
  /** Other */
  | "OTHER"
  /** Version 3 only. Picture book */
  | "PICTURE_BOOK"
  /** Video game */
  | "VIDEO_GAME"
  /** Video game driven primary by text and narrative */
  | "VISUAL_NOVEL"
  /** Version 3 only. Written works published online */
  | "WEB_NOVEL";

/** The current releasing status of the media */
export type MediaStatus =
  /** Ended before the work could be finished */
  | "CANCELLED"
  /** Has completed and is no longer being released */
  | "FINISHED"
  /** Version 2 only. Is currently paused from releasing and will resume at a later date */
  | "HIATUS"
  /** To be released at a later date */
  | "NOT_YET_RELEASED"
  /** Currently releasing */
  | "RELEASING";

export type AniListMediaFragment = {
  id: number;
  synonyms: Array<string | null> | null;
  siteUrl: string | null;
  bannerImage: string | null;
  episodes: number | null;
  format: MediaFormat | null;
  genres: Array<string | null> | null;
  season: MediaSeason | null;
  seasonYear: number | null;
  status: MediaStatus | null;
  averageScore: number | null;
  trending: number | null;
  popularity: number | null;
  favourites: number | null;
  duration: number | null;
  countryOfOrigin: string | null;
  source: MediaSource | null;
  isLicensed: boolean | null;
  isAdult: boolean | null;
  title: {
    romaji: string | null;
    english: string | null;
    native: string | null;
    userPreferred: string | null;
  } | null;
  coverImage: { extraLarge: string | null; large: string | null } | null;
  tags: Array<{ name: string; rank: number | null; isAdult: boolean | null } | null> | null;
  externalLinks: Array<{
    siteId: number | null;
    type: ExternalLinkType | null;
    isDisabled: boolean | null;
  } | null> | null;
  startDate: { year: number | null; month: number | null; day: number | null } | null;
  endDate: { year: number | null; month: number | null; day: number | null } | null;
  nextAiringEpisode: { episode: number; airingAt: number } | null;
  studios: { nodes: Array<{ name: string } | null> | null } | null;
};
// ...
export type CatalogQueryVariables = Exact<{
  page: number;
  sort?: Array<MediaSort | null | undefined> | MediaSort | null | undefined;
  search?: string | null | undefined;
  genre_in?: Array<string | null | undefined> | string | null | undefined;
  genre_not_in?: Array<string | null | undefined> | string | null | undefined;
  tag_in?: Array<string | null | undefined> | string | null | undefined;
  tag_not_in?: Array<string | null | undefined> | string | null | undefined;
  seasonYear?: number | null | undefined;
  season?: MediaSeason | null | undefined;
  format?: MediaFormat | null | undefined;
  status?: MediaStatus | null | undefined;
  licensedById?: number | null | undefined;
  countryOfOrigin?: string | null | undefined;
  source?: MediaSource | null | undefined;
  startDate_greater?: number | null | undefined;
  startDate_lesser?: number | null | undefined;
  episodes_greater?: number | null | undefined;
  episodes_lesser?: number | null | undefined;
  duration_greater?: number | null | undefined;
  duration_lesser?: number | null | undefined;
  isLicensed?: boolean | null | undefined;
}>;